pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "book-parser"
version = "1.0.0"
description = "LangChain multi-agent system for converting aptitude books into structured question banks"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "accelerate>=0.27.0",
    "qwen-vl-utils>=0.0.2",
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-huggingface>=0.0.3",
    "PyMuPDF>=1.24.0",
    "psycopg2-binary>=2.9.9",
    "pandas>=2.1.0",
    "Pillow>=10.0.0",
]

[project.scripts]
book-parser = "book_parser.main:main"
